// frontend/src/interfaces/Address.ts
export interface AddressInterface{
  ID?: number;
  HouseNo?:  string;    
	VillageNo?: string;    
	Village?:  string;   
	Alley?:     string;    
	Road?:  string; 
	SubDistrict?:  string;   
	District?: string;
	Province?: string;
	PostCode?: string;
  StudentID?: string; // เชื่อมโยงกับ Student
}