// frontend/src/interfaces/Other.ts
export interface OtherInteface{
  ID?: number;
  LatestGraduationFrom?: string;
  GraduatedYear?: number;
  Gpax?: number;
  PersonalVehicles?: string;
  Color?: string;
  PlateNo?: string;
  TaxDate?: string;
  ProvinceVehicle?: string;
  LicenseID?: number; 
  Type?: string;
  ExpiredCard?: string;
  StudentID?: string; // เชื่อมโยงกับ Student
}